+++ 4788/src/main/cpp/DriveSystem.cpp
@@ -20,6 +20,9 @@ DrivetrainManual::DrivetrainManual(std::string name,
   Requires(&drivetrain);
   SetCanBeInterrupted(true);
   SetCanBeReused(true);
+
+  _ChangeGears.SetTarget(wml::actuators::BinaryActuatorState::kForward); // Default State High Gear
+  _Shift2PTO.SetTarget(wml::actuators::BinaryActuatorState::kReverse); // Default State Offline
 }
 
 // On Loop Update, this code runs (Just a drivebase)
