+++ 4788/src/main/cpp/Turret.cpp
@@ -7,9 +7,9 @@ using namespace wml::controllers;
 Turret::Turret(Gearbox &Rotation,
  							 Gearbox &VerticalAxis, 
 							 Gearbox &FlyWheel, 
-							 sensors::LimitSwitch &LeftLimit, 
-							 sensors::LimitSwitch &RightLimit, 
-							 sensors::LimitSwitch &AngleDownLimit, 
+							 sensors::BinarySensor &LeftLimit, 
+							 sensors::BinarySensor &RightLimit, 
+							 sensors::BinarySensor &AngleDownLimit, 
 							 SmartControllerGroup &contGroup, 
 							 std::shared_ptr<nt::NetworkTable> &visionTable,
 							 std::shared_ptr<nt::NetworkTable> &rotationTable, 
@@ -36,84 +36,29 @@ Turret::Turret(Gearbox &Rotation,
 void Turret::ZeroTurret() {
 	// Zero Encoder on left limit
 	ZeroTimer.Start();
-	while (_LeftLimit.Get() < 1) {
-		if (ZeroTimer.Get() < ControlMap::TurretZeroTimeoutSeconds) {
-			_RotationalAxis.transmission->SetVoltage(12 * -0.3);
-		} else {
-			std::cout << "Turret Zero Timed Out" << std::endl;
-			_contGroup.GetController(ControlMap::CoDriver).SetRumble(wml::controllers::RumbleType::kLeftRumble, 1);
-			_contGroup.GetController(ControlMap::CoDriver).SetRumble(wml::controllers::RumbleType::kRightRumble, 1);
-			_RotationalAxis.transmission->SetVoltage(0);
-			break;
-		}
-	} 
-	_RotationalAxis.encoder->ZeroEncoder();
+
+	// Get Left Limit
+	Turret::TurretZeroLeft(ZeroTimer.Get());
+
+	// Minimum Rotations
 	MinRotation = (_RotationalAxis.encoder->GetEncoderRotations() + ControlMap::TurretEncoderSafeZone);
-	while (_RightLimit.Get() < 1) {
-		if (ZeroTimer.Get() < ControlMap::TurretZeroTimeoutSeconds) {
-			_RotationalAxis.transmission->SetVoltage(12 * 0.3);
-		} else {
-			std::cout << "Turret Zero Timed Out" << std::endl;
-			_contGroup.GetController(ControlMap::CoDriver).SetRumble(wml::controllers::RumbleType::kLeftRumble, 1);
-			_contGroup.GetController(ControlMap::CoDriver).SetRumble(wml::controllers::RumbleType::kRightRumble, 1);
-			_RotationalAxis.transmission->SetVoltage(0);
-			break;
-		}
-	}
+	
+	// Get Right Limit
+	Turret::TurretZeroRight(ZeroTimer.Get());
+
+	// Get Max Rotations
 	MaxRotation = (_RotationalAxis.encoder->GetEncoderTicks() - ControlMap::TurretEncoderSafeZone);
-	while(_AngleDownLimit.Get() < 1) {
-		if (ZeroTimer.Get() < ControlMap::TurretZeroTimeoutSeconds) {
-			_VerticalAxis.transmission->SetVoltage(12 * 0.2);
-		} else {
-			_VerticalAxis.transmission->SetVoltage(0);
-			_contGroup.GetController(ControlMap::CoDriver).SetRumble(wml::controllers::RumbleType::kLeftRumble, 1);
-			_contGroup.GetController(ControlMap::CoDriver).SetRumble(wml::controllers::RumbleType::kRightRumble, 1);
-			std::cout << "Turret Zero Timed Out" << std::endl;
-			break;
-		}
-	}
+
+	// Get Angle Limit
+	Turret::TurretZeroAngle(ZeroTimer.Get());
+	
+	// Reset Timers
 	ZeroTimer.Stop();
 	ZeroTimer.Reset();
-	_VerticalAxis.encoder->ZeroEncoder();
+
+	// Maxed Vertical Axis & Zero Flywheel
 	MaxAngleRotations = (_VerticalAxis.encoder->GetEncoderRotations() + ControlMap::TurretEncoderSafeZone);
 	_FlyWheel.encoder->ZeroEncoder();
-}
-
-
-
-double Turret::XAutoAimCalc(double dt, double targetx)  {
-
-	double TurretFullRotation = (ControlMap::TurretEncoderRotations * ControlMap::TurretRatio);
-	double Rotations2FOV = (TurretFullRotation/ControlMap::CamFOV);
-	double targetXinRotations = targetX * (Rotations2FOV/imageWidth);
-
-	Rgoal = _RotationalAxis.encoder->GetEncoderRotations() + targetXinRotations;
-	double input = _RotationalAxis.encoder->GetEncoderRotations();
-
-	// Calculate PID
-	Rerror = Rgoal - input;
-
-	double derror = (Rerror - RpreviousError) / dt;
-	Rsum = Rsum + Rerror * dt;
-
-	if (Rsum > (imageWidth/2)) {
-		Rsum = imageWidth;
-	} else if (Rsum < -(imageWidth/2)) {
-		Rsum = -imageWidth;
-	}
-
-	double output = RkP * Rerror + RkI * Rsum + RkD * derror;
-
-	// Convert to -1 - 1 for motor
-	output /= Rotations2FOV;
-
-	table->PutNumber("RoationDError", derror);
-	table->PutNumber("RotationError", Rerror);
-	table->PutNumber("RotationDelta Time", dt);
-	table->PutNumber("RotationOutput", output);
-
-	RpreviousError = Rerror;
-	return -output;
 }
 
 // Using Setpoints
@@ -122,29 +67,10 @@ double Turret::YAutoAimCalc(double dt, double TargetInput) {
 	double targetEncoderValue;
 	int LowPoint = 10;
 	int MaxPoint = 50;
-	int PixleMount = 2;
+	int PixleAmount = 2;
 
 	// Setpoint Selection.
-	targetEncoderValue = (TargetInput < LowPoint) && (TargetInput > LowPoint + PixleMount) ? ControlMap::AngleSetpoint1 : MaxPoint;
-	LowPoint += PixleMount;
-	targetEncoderValue = (TargetInput < LowPoint) && (TargetInput > LowPoint + PixleMount) ? ControlMap::AngleSetpoint2 : MaxPoint;
-	LowPoint += PixleMount;
-	targetEncoderValue = (TargetInput < LowPoint) && (TargetInput > LowPoint + PixleMount) ? ControlMap::AngleSetpoint3 : MaxPoint;
-	LowPoint += PixleMount;
-	targetEncoderValue = (TargetInput < LowPoint) && (TargetInput > LowPoint + PixleMount) ? ControlMap::AngleSetpoint4 : MaxPoint;
-	LowPoint += PixleMount;
-	targetEncoderValue = (TargetInput < LowPoint) && (TargetInput > LowPoint + PixleMount) ? ControlMap::AngleSetpoint5 : MaxPoint;
-	LowPoint += PixleMount;
-	targetEncoderValue = (TargetInput < LowPoint) && (TargetInput > LowPoint + PixleMount) ? ControlMap::AngleSetpoint6 : MaxPoint;
-	LowPoint += PixleMount;
-	targetEncoderValue = (TargetInput < LowPoint) && (TargetInput > LowPoint + PixleMount) ? ControlMap::AngleSetpoint7 : MaxPoint;
-	LowPoint += PixleMount;
-	targetEncoderValue = (TargetInput < LowPoint) && (TargetInput > LowPoint + PixleMount) ? ControlMap::AngleSetpoint8 : MaxPoint;
-	LowPoint += PixleMount;
-	targetEncoderValue = (TargetInput < LowPoint) && (TargetInput > LowPoint + PixleMount) ? ControlMap::AngleSetpoint9 : MaxPoint;
-	LowPoint += PixleMount;
-	targetEncoderValue = (TargetInput < LowPoint) && (TargetInput > LowPoint + PixleMount) ? ControlMap::AngleSetpoint10 : MaxPoint;
-	LowPoint += PixleMount;
+	targetEncoderValue = SetPointSelection(LowPoint, MaxPoint, PixleAmount, TargetInput);
 
 
 	// Calculate PID
@@ -166,6 +92,7 @@ double Turret::YAutoAimCalc(double dt, double TargetInput) {
 	// Convert to -1 - 1 for motor
 	output /= ControlMap::MaxAngleEncoderRotations;
 
+	// Temp Values
 	table->PutNumber("AngleDError", derror);
 	table->PutNumber("AngleError", Aerror);
 	table->PutNumber("AngleDelta Time", dt);
@@ -176,64 +103,15 @@ double Turret::YAutoAimCalc(double dt, double TargetInput) {
 	return output;
 }
 
-void Turret::TuneTurretPID() {
-
-	if (_contGroup.Get(ControlMap::kpUP, Controller::ONRISE)) {
-		RkP += 0.01;
-	} else if (_contGroup.Get(ControlMap::kpDOWN, Controller::ONRISE)) {
-		RkP -= 0.001;
-	} else if (_contGroup.Get(ControlMap::kiUP, Controller::ONRISE)) {
-		RkI += 0.001;
-	} else if (_contGroup.Get(ControlMap::kiDOWN, Controller::ONRISE)) {
-		RkI -= 0.001;
-	} else if (_contGroup.Get(ControlMap::kdUP, Controller::ONRISE)) {
-		RkD += 0.001;
-	}	else if (_contGroup.Get(ControlMap::kdDOWN, Controller::ONRISE)) {
-		RkD -= 0.001;
-	}
-
-	table->PutNumber("kP", RkP);
-	table->PutNumber("kI", RkI);
-	table->PutNumber("kD", RkD);
-	table->PutNumber("Sum", Rsum);
-}
-
-void Turret::TuneAnglePID() {
-
-	if (_contGroup.Get(ControlMap::kpUP, Controller::ONRISE)) {
-		AkP += 0.01;
-	} else if (_contGroup.Get(ControlMap::kpDOWN, Controller::ONRISE)) {
-		AkP -= 0.001;
-	} else if (_contGroup.Get(ControlMap::kiUP, Controller::ONRISE)) {
-		AkI += 0.001;
-	} else if (_contGroup.Get(ControlMap::kiDOWN, Controller::ONRISE)) {
-		AkI -= 0.001;
-	} else if (_contGroup.Get(ControlMap::kdUP, Controller::ONRISE)) {
-		AkD += 0.001;
-	}	else if (_contGroup.Get(ControlMap::kdDOWN, Controller::ONRISE)) {
-		AkD -= 0.001;
-	}
-
-	table->PutNumber("kP", AkP);
-	table->PutNumber("kI", AkI);
-	table->PutNumber("kD", AkD);
-	table->PutNumber("Sum", Asum);
-}
-
 void Turret::TeleopOnUpdate(double dt) {
 
 	targetX = table->GetNumber("Target_X", 0)/imageWidth;
 	targetY = table->GetNumber("Target_Y", 0)/imageHeight;
 
-
-	if (ControlMap::TuneTurretPID && ControlMap::TuneAnglePID) {
-		std::cout << "Conflict Turret/Angle Tuning" << std::endl;
-	}	else if (ControlMap::TuneTurretPID) {
-		TuneTurretPID();
-	} else if (ControlMap::TuneAnglePID) {
-		TuneAnglePID();
-	}
+	// Tune Turret PID (If active)
+	PIDTuner();
 	
+
 	if (_contGroup.Get(ControlMap::TurretAutoAim)) {
 		if (targetX > imageWidth || targetY > imageHeight) {
 			std::cout << "Error: Target is artifacting" << std::endl;
@@ -245,7 +123,6 @@ void Turret::TeleopOnUpdate(double dt) {
 
 
 	if (!_TurretDisable) {
-
 		// Manual Angle Control
 		AngularPower += std::fabs(_contGroup.Get(ControlMap::TurretManualAngle)) > ControlMap::joyDeadzone ? _contGroup.Get(ControlMap::TurretManualAngle) : 0;
 
@@ -263,13 +140,7 @@ void Turret::TeleopOnUpdate(double dt) {
 	}
 
 	// Flywheel Feedback
-	if (_FlyWheel.encoder->GetEncoderAngularVelocity() >= ControlMap::FlyWheelVelocity) {
-		_contGroup.GetController(ControlMap::CoDriver).SetRumble(wml::controllers::RumbleType::kLeftRumble, 1);
-		_contGroup.GetController(ControlMap::CoDriver).SetRumble(wml::controllers::RumbleType::kRightRumble, 1);
-	} else {
-		_contGroup.GetController(ControlMap::CoDriver).SetRumble(wml::controllers::RumbleType::kLeftRumble, 0);
-		_contGroup.GetController(ControlMap::CoDriver).SetRumble(wml::controllers::RumbleType::kRightRumble, 0);
-	}
+	ContFlywheelFeedback();
 
 	// Limits Turret Speed
 	RotationPower *= ControlMap::MaxTurretSpeed;
